add: adicionando a classe ExceptionManager e outras correcoes

+++ api-produtos/src/main/java/br/com/oaktecnologia/entity/dto/ProductInfoDto.java
@@ -2,7 +2,9 @@ package br.com.oaktecnologia.entity.dto;
 
 import br.com.oaktecnologia.entity.Product;
 
-public record ProductInfoDto(Long id, String name, String description, Double price) {
+import java.math.BigDecimal;
+
+public record ProductInfoDto(Long id, String name, String description, BigDecimal price) {
     public ProductInfoDto(Product product) {
         this(product.getId(), product.getName(), product.getDescription(), product.getPrice());
     }
